Refactor editing
